hw3/booking-service/requirements.txt
grpcio==1.59.2
grpcio-tools==1.59.2
protobuf==4.25.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]
tenacity==8.2.3